part

+++ app/src/main/java/com/example/onlinestore/domain/models/AllCategory.kt
@@ -6,6 +6,6 @@ data class AllCategory(
     val id: String,
     val title: String,
     val descripition: String,
-    val prise: Float,
+    val prise: Int,
     val imageUrl: String,
 )
